+++ example/lib/main.dart
@@ -1,17 +1,22 @@
 import 'package:flutter/material.dart';
+
 import 'package:flutter_native_contact_picker/flutter_native_contact_picker.dart';
+import 'package:flutter_native_contact_picker/model/contact.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
-  final FlutterContactPicker _contactPicker = new FlutterContactPicker();
+  final FlutterNativeContactPicker _contactPicker =
+      FlutterNativeContactPicker();
   List<Contact>? _contacts;
 
   @override
@@ -21,18 +26,18 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new Scaffold(
-        appBar: new AppBar(
-          title: new Text('Contact Picker Example App'),
+    return MaterialApp(
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('Contact Picker Example App'),
         ),
-        body: new Center(
-          child: new Column(
+        body: Center(
+          child: Column(
             mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              new MaterialButton(
+              MaterialButton(
                 color: Colors.blue,
-                child: new Text("Single"),
+                child: const Text("Single"),
                 onPressed: () async {
                   Contact? contact = await _contactPicker.selectContact();
                   setState(() {
@@ -40,9 +45,9 @@ class _MyAppState extends State<MyApp> {
                   });
                 },
               ),
-              new MaterialButton(
+              MaterialButton(
                 color: Colors.blue,
-                child: new Text("Multiple"),
+                child: const Text("Multiple"),
                 onPressed: () async {
                   final contacts = await _contactPicker.selectContacts();
                   setState(() {