+++ test/flutter_native_contact_picker_method_channel_test.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+// import 'package:flutter_native_contact_picker/flutter_native_contact_picker_method_channel.dart';
+
+void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  // MethodChannelFlutterNativeContactPicker platform = MethodChannelFlutterNativeContactPicker();
+  const MethodChannel channel = MethodChannel('flutter_native_contact_picker');
+
+  setUp(() {
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
+      channel,
+      (MethodCall methodCall) async {
+        return '42';
+      },
+    );
+  });
+
+  tearDown(() {
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
+  });
+
+  test('getPlatformVersion', () async {
+    // expect(await platform.getPlatformVersion(), '42');
+  });
+}